+++ .bowerrc
@@ -1,0 +1,3 @@
+{
+    "directory": "public/bower_components"
+}
