+++ public/app/restaurants.html
@@ -1,0 +1,29 @@
+<div class="span6">
+    <table class="table table-striped table-condensed">
+        <thead>
+            <tr>
+                <th>Name</th>
+                <th>Address</th>
+                <th>City</th>
+                <th>State</th>
+                <th>Zipcode</th>
+                <th>Cuisine</th>
+                <th>Type</th>
+                <th>Price</th>
+            </tr>
+        </thead>
+        <tbody>
+            <tr ng-repeat="restaurant in restaurants">
+                <td>{{ restaurant.name }}</td>
+                <td>{{ restaurant.address }}</td>
+                <td>{{ restaurant.city }}</td>
+                <td>{{ restaurant.state }}</td>
+                <td>{{ restaurant.zipcode }}</td>
+                <td>{{ restaurant.cuisine }}</td>
+                <td>{{ restaurant.type }}</td>
+                <td>{{ restaurant.price }}</td>
+            </tr>
+        </tbody>
+    </table>
+    <a ng-click="createNewRestaurant()" class="btn btn-small">Add New Restaurant</a> 
+</div>
