+++ public/app/services.js
@@ -1,0 +1,33 @@
+var services = angular.module('restaurants.services', ['ngResource']);
+
+services.factory('RestaurantsFactory', function($resource){
+    return $resource('/api/restaurants', {}, {
+        query: {
+            method: 'GET',
+            isArray: true
+        },
+        create: {
+            method: 'POST'
+        }
+    })
+});
+
+services.factory('RestaurantFactory', function($resouce){
+    return $resource('/api/restaurant/:id', {}, {
+        show: {
+            method: 'GET'
+        },
+        update: {
+            method: 'PUT',
+            params: {
+                id: '@id'
+            }
+        },
+        delete: {
+            method: 'DELETE',
+            params: {
+                id: '@id'
+            }
+        }
+    })
+});
