+++ src/main/java/com/theironyard/charlotte/controller/RestaurantController.java
@@ -1,5 +1,7 @@
 package com.theironyard.charlotte.controller;
 
+import org.springframework.beans.factory.annotation.Configurable;
+import org.springframework.boot.context.embedded.ConfigurableEmbeddedServletContainer;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -10,9 +12,9 @@ import org.springframework.web.bind.annotation.RequestMethod;
  */
 @Controller
 public class RestaurantController {
-
+    
     @RequestMapping(path = "/", method = RequestMethod.GET)
-    public String home(Model model) {
+    public String home(Model model, ConfigurableEmbeddedServletContainer container) {
         return "home";
     }
 }
