+++ src/main/resources/templates/home.html
@@ -3,13 +3,21 @@
 <head>
     <meta charset="UTF-8">
     <title>Home</title>
-    <script src="/public/app.js" />
+    
 </head>
-<body ng-app="restaurants" ng-strict-di>
+<body ng-app="restaurants">
 
 <div class="idk" ui-view="restaurants">
 
 </div>
+<script src="/bower_components/angular/angular.js"></script>
+<script src="/bower_components/angular-resource/angular-resource.js"></script>
+<script src="/bower_components/angular-route/angular-route.js"></script>
+<script src="/app/controller.js"></script>
+<script src="/app/services.js"></script>
+<script src="/app/app.js"></script>
+
+
 
 </body>
 </html>
